Initial commit
Switchgear is a Java library that provides an API for call isolation, timeouts,
and circuit breaker functionality. Our main goal is to achieve minimal performance
overhead of the library, even when used in the hottest place of your application.
Additionally, we've opted for zero transitive dependencies and maintaining a fluent API.

Switchgear is used internally in Smaato for several years supporting OpenRTB auction.
Whis this commit it is being moved to a public domain.